age: Split up parsing of age-native and SSH identities
Parsing of age-native identities is moved to a new IdentityFile struct,
and age::cli_common::read_identities now returns Vec<Box<dyn Identity>>,
as under the hood it now contains both age::IdentityFile and
age::ssh::Identity, which both implement age::Identity.
