+++ age/CHANGELOG.md
@@ -12,6 +12,8 @@ to 1.0.0 are beta releases.
 ### Added
 - `age::Identity` trait, representing an identity that can decrypt an age file.
   All relevant `age` types implement this trait.
+- `age::IdentityFile` struct, for parsing a list of native age identities from a
+  file.
 - Asynchronous APIs for encryption and decryption, enabled by the `async`
   feature flag:
   - `age::Encryptor::wrap_async_output()`
@@ -37,6 +39,12 @@ to 1.0.0 are beta releases.
 - `age::Format` has been moved to `age::armor::Format`.
 - SSH support has been moved into the `age::ssh` module.
 - OpenSSH `ssh-rsa` keys are now supported without the `unstable` feature flag.
+- `age::cli_common::read_identities` now returns `Vec<Box<dyn Identity>>`, as it
+  abstracts over `age::IdentityFile` and `age::ssh::Identity`.
+
+### Removed
+- `age::keys::{Identity, IdentityKey}` (replaced by `age::Identity` trait on
+  individual identities, and `age::IdentityFile` for parsing identities).
 
 ## [0.4.0] - 2020-03-25
 ### Added
