+++ rage/src/bin/rage/main.rs
@@ -354,12 +354,7 @@ fn decrypt(opts: AgeOptions) -> Result<(), error::DecryptError> {
             )?;
 
             decryptor
-                .decrypt_with_callbacks(
-                    identities
-                        .into_iter()
-                        .map(|i| Box::new(i) as Box<dyn age::Identity>),
-                    &UiCallbacks,
-                )
+                .decrypt_with_callbacks(identities.into_iter(), &UiCallbacks)
                 .map_err(|e| e.into())
                 .and_then(|input| write_output(input, output))
         }
